Move the column specific pad calculator tests into separate files
